pyproject.toml
[tool.poetry]
name = "nautilus_mt5"
version = "0.1.7"
description = "Nautilus Trader MetaTrader5 Adapter"
readme = "README.md"
repository = "https://github.com/quantspub/nautilus_mt5"
authors = ["Quants Pub <quantspub@gmail.com>"]
include = ["nautilus_mt5/**/*"]
packages = [{ include = "nautilus_mt5" }]

[build-system]
requires = [
    "setuptools>=64",
    "poetry-core>=1.9.0",
    "toml>=0.10.2",
    "numpy>=1.26.4",
    "Cython<=3.1.0a1",
]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
cython = "<=3.1.0a1"
numpy = ">=1.21.6, <=1.26.4"
nautilus-trader = "^1.206.0"
docker = "^7.1.0"
python-dotenv = "^1.0.1"
numba = "^0.60.0"
pyparsing = "^3.1.2"
toml = "^0.10.2"
metatrader5ext = {path = "../metatrader5ext"}

[tool.poetry.group.test]
optional = true

[tool.poetry.group.test.dependencies]
pytest = "^8.3.2"
